[move-function] Add a new pass that propagates dominating debug_value into coroutine func-lets.
The big picture here is that I am trying to transition IRGen to use
llvm.dbg.addr instead of llvm.dbg.declare. This will let us model move semantics
in the debugger by using inserting debug_value undef at the SIL level (which
invalidates the value). In a recent patch when I tried to perform this, we
discovered that everything worked... except for the debug info around variables
in async functions. This commit is a collaboration in between Adrian Prantl
(@aprantl) and I. Thanks to AdrianP for his help here!

Big picture is that there is a bunch of work that is done in the LLVM level in
the coroutine splitter to work around communicating live variables in the
various coroutine func-lets. This logic is all done with debug.declare and we
would need to update that logic in the coroutine splitter to handle
debug.addr. Rather than do this, after some conversation, AdrianP and I realized
that we could get the same effect of a debug.declare by just redeclaring the
current live set of debug_value after each possible coroutine funclet
start. This in practice this as an algorithm works as follows:

1. We walk the CFG along successors. By doing this we guarantee that we visit
   blocks after their dominators.

2. When we visit a block, we walk the block from start->end. During this walk:

   a. We grab a new block state from the centralized block->blockState map. This
      state is a [SILDebugVariable : DebugValueInst].

   b. If we see a debug_value, we map blockState[debug_value.getDbgVar()] =
      debug_value. This ensures that when we get to the bottom of the block, we
      have pairs of SILDebugVariable + last debug_value on it.

   c. If we see any coroutine funclet boundaries, we clone the current tracked
      set of our block state and then walk up the dom tree dumping in each block
      any debug_value with a SILDebugVariable that we have not already
      dumped. This is maintained by using a visited set of SILDebugVariable for
      each funclet boundary.

The end result is that at the beginning of each funclet we will basically
declare the debug info for an addr.