Override -arch For Stdlib Targets
On Apple Silicon hosts, CMake defaults OSX_ARCHITECTURES to the
host arch. When the cmake later explicitly provides -arch, clang
interprets this as a directive to emit fat object files. When we go
to lipo them together, the host arches in the object files collide
and the build goes kaput.