ClangImporter: Import 'typedef id<P, Q> Foo' as a constraint type rather than an existential type
We want to be able to use these typealiases both as existential types
and as generic constraints, so unwrap the ExistentialType explicitly
before setting the underlying type of the typealias.

This might not be the best long-term approach, so I added a FIXME
comment with my thoughts there.

Fixes rdar://problem/88208893.