Use the correct spare bits mask; allow MPE record to grow in the future
The runtime wants the mask of "common spare bits for the
enum payloads", which we can generate inside the compiler
by AND-ing together the spare bits for each payload into
a single mask.

To allow the MPE record to resize in the future, express everything
as indexes in a "contents" array (of 32-bit values) with the size of
that array being part of the first element.  I've commented the
boilerplate for adding a new field (computing the index of the
added property, returning default value if the record isn't big enough).

With this, I'm finally passing my current tests, and have some local
validation that the compiler-provided spare bits mask in the
reflection record matches exactly the spare bits mask that the
runtime can compute from first principles.  This in turn gives
us some confidence that the runtime will be able to properly
reflect enum payloads for older binaries that lack the new reflection
record.