Runtime: make padding explicit for `ElementStorage`
The `ElementStorage` member `Capacity` will be padded up to the next pointer alignment due to the storage of the `ElemTy` by C's struct layout rules.  However, this is implicit and not entirely guaranteed.  Instead, make the storage a pointer sized value, and then truncate to 32-bits to maintain ABI compatibility.  This becomes important as we expand on the runtime metadata handling.