GSB: Fix null pointer dereference in involvesNonSelfSubjectTypes()
If we have a RequirementSource here that's not rooted in a
RequirementSignatureSelf, we would dereference a null pointer
instead of ending the loop early.

The only way such a RequirementSource can appear is from
inferConditionalRequirements().

This crash does not occur on the main branch, because
inferConditionalRequirements() is no longer performed inside
protocols there because the RequirementMachine can't support it.

A better fix would be to change inferConditionalRequirements() to
build the proper RequirementSource in this case, but I'm going
with a narrow fix since the GenericSignatureBuilder is going away
on the main branch at some point.

Fixes <rdar://problem/88474300> and <https://bugs.swift.org/browse/SR-15792>.