RequirementMachine: Refine elimination order for concrete type requirements
Skipping the check if either rule was not a concrete type rule is
not correct. For example, if we had these two elimination candidates:

    X.T.[concrete: G<Int>] => X.T
    X.U.[concrete: Int] => X.U

We would correctly pick the first one for elimination. However, if
another non-concrete type rule appeared in the middle, such as

    X.T.[concrete: G<Int>] => X.T
    X.Z => X.Y
    ...
    X.U.[concrete: Int] => X.U

We would pick 'X.Z => X.Y' as the candidate, and then pick
'X.U.[concrete: Int] => X.U' upon encountering it.

Fix this so that we have a true linear order here.