[5.6][swift-stdlib-tool] Fix finding sibling directories
When used without passing `--source-libraries`, this tool fetches its
current path and uses that to find the directories of the swift
libraries near it.

Since https://github.com/apple/swift/pull/39216 it was searching the
nearby directories recursively, and only for files, instead of just
looking at 1 level, only for directories. Because of this it would never
discover `path/to/lib/swift-5.0/...` because it was only looking for
files.

This fixes this issue with a new function, and renames the previous
function to be more explicit for the other use case that does want that
behavior.

(cherry picked from commit d1eabf182c7c1abf36b28b111ec98c0ca5168efe)