[SIL] Make owned function arguments lexical.
The destroys of owned arguments must not be hoisted over deinit barriers
to respect the semantics of lexical lifetimes.  Here, owned SILValues
which are SILFunctionArguments are made to be lexical.  Code which
hoists destroys will check this field and not hoist destroys.

Allowed more RAUWing to be done with lexical values; specifically,
allowed the replacement of a non-lexical value with a lexical value if
the replacing value's lifetime wouldn't be extended as a result.

Updated tests to get owned values via applies rather than having them
passed as arguments.