RequirementMachine: Simpler handling of concrete conformances in minimal conformances algorithm
This logic was tricky and unsound.

Correct handling of concrete conformances requires rebuilding the
signature after dropping protocol conformance requiremnts made
redundant by concrete type requirements.

The old support for "derived-via-concrete" requirements in the
minimal conformances algorithm was causing other problems, and it
is the wrong approach anyway, so just remove it.