Expose MPE Spare bit mask as a global var in target, fetch from host
This allows the host to query the target directly for the appropriate
spare bit mask info.  It requires the memory reader interface to
support `getSymbolAddress` but should otherwise be fully general.

Note: The growing stack of closures in TypeRefBuilder used to insulate this
non-template type from the template types it uses is getting pretty cumbersome.
Surely there's a better design?