[debug-info] Begin caching SILDebugVariable in SILModule.
SILDebugVariable is an originally small value type that has grown over time to
120 bytes. This is rather large to be manipulating as a value type just from an
efficiency perspective. In this two part commit, I refactor the code base so
that SILDebugVariable becomes an immutable value type that is cached on
SILModule. The reason why we keep it as a value type vs vending a pointer is
that values of these types are often times passed around as
Optional<SILDebugVariable>. Changing this would require a large code base churn
and require us ot use Optional<SILDebugVariable *>, SILDebugVariable *, or
NullablePtr<SILDebugVariable>... which would be unfortunate.

In order to avoid this churn in the code base (and loss of safety), this PR:

1. Keep SILDebugVariable as a value type, but have it internally contain a
   pointer to a Storage struct that is what is actually cached on the
   SILModule. This additionally gives us simple a tombstone and empty
   representation for SILDebugVariable since we can make our storage pointer
   UINT_MAX and UINT_MAX-1 sentinels allowing this to be used in DenseMaps.

2. Hides the constructors of SILDebugVariable and make it so that one can only
   create one by calling the static method SILDebugVariable::get. That method
   then uses a folding set on the passed in module argument to look up the
   appropriate unique storage (or create one) and then return the storage inside
   a wrapped SILDebugVariable.

3. Since SILDebugVariable is now immutable, hide its fields, introduce only
   getters (not setters), and introduce a Builder struct
   (SILDebugVariable::Builder) that lets one build up a new SILDebugVariable
   based off of either a null representation or by templating on an existing
   SILDebugVariable. When the builder is finalized, it just calls
   SILDebugVariable::get for you.

In this commit, I make the above changes. In the next commit in order to make it
easy to review, I update the rest of the codebase using these constructs. This
was done to make it easier to review the code. Fundamentally though the update
was mechanical since all of the constructs I changed was mechanical.

rdar://83957028