Rather than turning off sil-verify-all on all stdlib builds... Re-enable it only on the macOS x86_64 platform.
Most of the stdlibs at the SIL level are the same so by not verifying them all,
we aren't losing that much coverage. This gives us back some of the coverage we
lost when we disabled -sil-verify-all everywhere without causing us to have the
huge slow down when building multiple stdlibs.