Remove disable-objc-interop from test hermetic-seal-exec.swift
This test uses the stdlib build from the current build which typically
will have objc interop enabled.

If disable-objc-interop is enabled we make different layout assumptions
so mixing build products with the flag enabled/disabled is not supported.

This will get exposed when by running the test in optimize mode after PR #41338.

rdar://88827015