[move-function] When emitting debug_value undef, use the loc of the original DbgVarCarryingInst we found.
LLVM wants the debug_value undef to be /exactly/ the same as the original
debug_value including the loc. If we don't match, LLVM won't invalidate the
value.

Also now that things actually work (at least for copyable lets), lets add some
basic tests that debug info works as expected. NOTE: With this patch it seems
that address only move is still broken. I am looking into it and will fix it in
a forthcoming PR.